% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydflood.R
\docType{package}
\name{hydflood}
\alias{hydflood}
\alias{hydflood-package}
\title{hydflood: Flood Extents and Durations along the Rivers Elbe and Rhine}
\description{
Raster based flood modelling internally using \pkg{hyd1d}, an
R package to interpolate 1d water level and gauging data. The package
computes flood extent and durations through strategies originally developed
for 'INFORM', an 'ArcGIS'-based hydro-ecological modelling framework. It
does not provide a full, physical hydraulic modelling algorithm, but a
simplified, near real time 'GIS' approach for flood extent and duration
modelling. Computationally demanding annual flood durations have been
computed already and data products were published by Weber (2022)
\url{doi:10.1594/PANGAEA.948042}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hydflood.bafg.de}
  \item \url{https://github.com/bafg-bund/hydflood}
  \item Report bugs at \url{https://github.com/bafg-bund/hydflood/issues/}
}

}
\author{
\strong{Maintainer}: Arnd Weber \email{arnd.weber@bafg.de} (\href{https://orcid.org/0000-0002-5973-2770}{ORCID})

Other contributors:
\itemize{
  \item Stephan Rosenzweig \email{rosenzweig@bafg.de} [contributor]
  \item Benjamin Eberhardt \email{eberhardt@bafg.de} [contributor]
}

}
