% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sf.af}
\alias{sf.af}
\title{Obtain projected versions of \code{sf.afe} and \code{sf.afr}}
\usage{
sf.af(name = NULL)
}
\arguments{
\item{name}{either 'Elbe' or 'Rhine'.}
}
\value{
\code{sf} with the projected active floodplain
}
\description{
Obtain projected versions of \code{\link{sf.afe}} and
\code{\link{sf.afr}}
}
\examples{
  library(hydflood)
  sf.af(name = "Elbe")

}
\seealso{
\code{\link{sf.afe}}, \code{\link{sf.afr}}
}
