% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sf.tiles}
\alias{sf.tiles}
\title{Obtain projected versions of \code{sf.tiles_elbe} and
\code{sf.tiles_rhine}}
\usage{
sf.tiles(name = NULL)
}
\arguments{
\item{name}{either 'Elbe' or 'Rhine'.}
}
\value{
\code{sf} with projected tiles
}
\description{
Obtain projected versions of \code{\link{sf.tiles_elbe}} and
\code{\link{sf.tiles_rhine}}
}
\examples{
  library(hydflood)
  sf.tiles(name = "Elbe")

}
\seealso{
\code{\link{sf.tiles_elbe}}, \code{\link{sf.tiles_rhine}}
}
