% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydflood.R
\docType{package}
\name{hydflood}
\alias{hydflood}
\alias{hydflood-package}
\title{hydflood: Flood Extents and Durations along the Rivers Elbe and Rhine}
\description{
Raster based flood modelling internally using \pkg{hyd1d}, an
R package to interpolate 1d water level and gauging data. The package
computes flood extent and durations through strategies originally developed
for 'INFORM', an 'ArcGIS'-based hydro-ecological modelling framework. It
does not provide a full, physical hydraulic modelling algorithm, but a
simplified, near real time 'GIS' approach for flood extent and duration
modelling. Computationally demanding annual flood durations have been
computed already and data products were published by Weber (2022)
\url{doi:10.1594/PANGAEA.948042}.
}
